detrModel/requirements.txt
watchdog
transformers
torch
Pillow